+++ common/src/main/java/ddraig/net/entropica/block/VoidStalkerOrchidBlock.java
@@ -17,10 +17,11 @@ import net.minecraft.world.level.block.BonemealableBlock;
 import net.minecraft.world.level.block.BushBlock;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class VoidStalkerOrchidBlock extends BushBlock implements BonemealableBlock {
+public class VoidStalkerOrchidBlock extends EntropicaFlowerBlock {
     public VoidStalkerOrchidBlock(Properties properties) {
         super(properties);
     }
+
 
     @Override
     protected boolean mayPlaceOn(BlockState state, BlockGetter level, BlockPos pos) {